\docType{package}
\name{rustfare}
\alias{rustfare}
\alias{rustfare-package}
\title{rustfare: toolbox for R language for open welfare data from Russian Federation}
\description{
  The rustfare package pools together tools for searching,
  extracting, and formatting welfare data on Russian
  Federation from different data providers. The rustfare
  functions are named according to data provider or the
  data set. Currently there data available from: Rosstat
  Regional Statistics, OECD, Quality of Governance
  -institute and World Bank
}
\details{
  rustfare: toolbox for R language for open welfare data
  from Russian Federation

  \tabular{ll}{ Package: \tab rustfare\cr Type: \tab
  Package\cr Version: \tab See sessionInfo() or DESCRIPTION
  file\cr Date: \tab 2013\cr License: \tab MIT\cr LazyLoad:
  \tab yes\cr }
}
\examples{
# library(rustfare)
}
\author{
  Markus Kainu <markuskainu(at)gmail.com>
  \url{http://markuskainu.fi/rustfare}
  \url{https://github.com/muuankarski/rustfare}
}

